VR: move files to their correct folders
