+++ lib/ValueRelations/Relations.cpp
@@ -1,4 +1,4 @@
-#include "Relations.h"
+#include "dg/ValueRelations/Relations.h"
 
 #include <algorithm>
 #include <cassert>
