+++ tools/CMakeLists.txt
@@ -119,11 +119,10 @@ endif (HAVE_SVF)
                                 PRIVATE dgllvmslicer
 				PRIVATE ${llvm_irreader})
 
-	add_executable(llvm-vr-dump llvm-vr-dump.cpp
-								${CMAKE_SOURCE_DIR}/include/dg/llvm/ValueRelations/Relations.cpp
-								${CMAKE_SOURCE_DIR}/include/dg/llvm/ValueRelations/VR.cpp)
+	add_executable(llvm-vr-dump llvm-vr-dump.cpp)
 	target_link_libraries(llvm-vr-dump
-				# dynamic LLVM
+				PRIVATE dganalysis
+				PRIVATE dgllvmvra
 				PRIVATE ${llvm}
 				# static LLVM
 				PRIVATE ${llvm_irreader})
